float

+++ src/Box2D.h
@@ -12,25 +12,25 @@ class Box2D
 {
 private:
     /* Инициализация координат прямоугольника*/
-    double box2d_x1 = 0.0,
+    float box2d_x1 = 0.0,
                  box2d_y1 = 0.0,
                  box2d_x2 = 0.0,
                  box2d_y2 = 0.0;
     /* Инициализация площади */
-    double box2d_square = 0.0;
+    float box2d_square = 0.0;
 
 public:
 
     Box2D() = default;
 
     /* Иницилизация координат центра*/
-    double box2d_x_center = 0.0,
+    float box2d_x_center = 0.0,
                  box2d_y_center = 0.0;
     /* Длина и высота */
-    double box2d_lenght = 0.0,
+    float box2d_lenght = 0.0,
                  box2d_height = 0.0;
     /* Координаты углов*/
-    double x1 = 0.0, y1 = 0.0, x2 = 0.0, y2 = 0.0;
+    float x1 = 0.0, y1 = 0.0, x2 = 0.0, y2 = 0.0;
 
     /* Обычный конструктор с параметрами*/
     Box2D(double x1, double y1, double x2,double y2)
@@ -91,8 +91,8 @@ void boxCrossing( Box2D&obj1, Box2D&obj2 )
   obj1.boxGeometry();
   obj2.boxGeometry();
 
-    double rho_x = 0.0, rho_y = 0.0;
-    double l = 0.0, h = 0.0;
+    float rho_x = 0.0, rho_y = 0.0;
+    float l = 0.0, h = 0.0;
 
 /* Вычисление растояния между центрами по осям x и y */
 rho_x =fabs(obj1.box2d_x_center - obj2.box2d_x_center);
